refactor: 클래스 위치 변경 (패키지 변경)
