+++ src/main/java/nextstep/security/domain/MemberDetailService.java
@@ -1,4 +1,4 @@
-package nextstep.security;
+package nextstep.security.domain;
 
 public interface MemberDetailService {
     MemberDetail findByUsername(String username);
