+++ src/main/java/nextstep/security/exception/AuthenticationException.java
@@ -1,4 +1,4 @@
-package nextstep.security;
+package nextstep.security.exception;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
